WIP search: use an enum for repo status fields
This more directly represents the different states a repository can be
in. This change was motivated by making it easier to communicate
progress updates. Additionally we avoid the need for our `dedupSort`
call when creating a list of resolvers.

This is likely overkill, but I got started exploring it and I quite like
how it looks. I have yet to update tests, but WDYT?