fix issue where scrollbars appear during resizing
The Resizable element previously used a "ghost" element to make it so that resizing would continue (and the `mousemove` event would keep firing) even if the user's cursor moved beyond the element's bounds (which occurs because the user moves the mouse faster than the element resizes, by definition, because the latter is in response to the former). This ghost element, however, causes scrollbars to appear because it enlarges the parent element. It is possible to suppress these through complex CSS, but that's error prone because the Resizable component is used in many different places.

Instead, we solve the problem by adding the `mousemove` listener to the root DOM node, so it will continue to fire.